Make authentication logic into custom hook
